requirements.txt
django
djangorestframework
gunicorn
dj-database-url
psycopg2-binary
python-dotenv